added api error handler

+++ controllers/team.js
@@ -2,70 +2,68 @@ const team = require('../models/team');
 const user = require('../models/user');
 const teamList = require('../models/teamList');
 const hash = require('../services/bcrypt');
+const ApiError = require('../utils/ApiErrors');
 
-exports.getPlayersAndTeam = async(req,res)=>{
+exports.getPlayersAndTeam = async(req,res,next)=>{
     try{
-        const {email ,password} = req.body;
-        const getTeam = await team.findOne({where:{email:email},attributes:['id','name','password']});
+        const {name ,password} = req.body;
+        if(isNullValue(name) || isNullValue(password))throw new ApiError('invalid input!' ,400)
+        const getTeam = await team.findOne({where:{name:name},attributes:['id','name','password']});
         if(getTeam){
-            console.log(email,password,getTeam.password)
             await hash.compareHash(password, getTeam.password);
-            console.log(email,password)
-            const getplayersInTeam = await getTeam.getUsers({attributes:['id','name','email']});
-            const teamName = await team.findOne({where:{email:email},attributes:['id','name']});
+            const getplayersInTeam = await getTeam.getUsers({attributes:['id','name']});
+            const teamName = await team.findOne({where:{name:name},attributes:['id','name']});
             res.status(200).json({success:true ,players:getplayersInTeam ,team:teamName});
         }else{
             res.status(404).json({success:false ,message:'team not found'});
         }
     }catch(err){
-        console.log(err)
-        res.status(500).json({success:false ,message:err.message});
+        next(new ApiError(err.message ,err.statusCode))
     }
 }
-exports.getPlayers = async(req,res)=>{
+exports.getPlayers = async(req,res,next)=>{
     try{
         const {id} = req.params;
-        const getTeam = await team.findOne({where:{id:id},attributes:['id','name']});
-        const getplayersInTeam = await getTeam.getUsers({attributes:['id','name','email']});
-        
-        res.status(200).json({success:true ,players:getplayersInTeam ,team:getTeam});
+        if(isNullValue(id))throw new ApiError('invalid input!' ,400)
+        const getTeam = await team.findOne({where:{id:id},attributes:['id','name','captainName']});
+        const getplayersInTeam = await getTeam.getUsers({attributes:['id','name']});
+        const isCaptain = getTeam.captainName === req.user.name?true:false;
+        res.status(200).json({success:true ,players:getplayersInTeam ,team:getTeam ,isCaptain});
     }catch(err){
-        console.log(err)
-        res.status(500).json({success:false ,message:err.message});
+        next(new ApiError(err.message ,err.statusCode))
     }
 }
 
-exports.getTeam = async(req,res)=>{
+exports.getTeam = async(req,res,next)=>{
     try {
         const getUser = req.user;
-        const team = await getUser.getTeams({attributes:['id','name']})
-        res.status(200).json({success:true ,team});
+        const playersInTeam = await getUser.getTeams({attributes:['id','name']})
+        res.status(200).json({success:true ,team:playersInTeam});
     } catch (err) {
-        res.status(500).json({success:false ,message:err.message});
+        next(new ApiError(err.message ,err.statusCode))
     }
 }
 
-exports.postTeam = async(req,res)=>{
+exports.postTeam = async(req,res,next)=>{
     try{
         const {name ,password} = req.body;
-        if(!password)throw new Error('please enter password')
+        if(isNullValue(name) || isNullValue(password))throw new ApiError('invalid input!' ,400)
         const hashPassword = await hash.createHash(password);
-        const post = await team.create({name ,email:req.user.email ,password :hashPassword});
+        const post = await team.create({name ,captainName:req.user.name ,password :hashPassword});
         await teamList.create({
             userId:req.user.id,
             teamId:post.id
         });
         res.status(200).json({success:true,team:post});
     }catch(err){
-        console.log(err)
-        res.status(500).json({success:false});
+        next(new ApiError(err.message ,err.statusCode))
     }
 }
-exports.deleteTeam = async(req,res)=>{
+exports.deleteTeam = async(req,res,next)=>{
     try{
         const {id} = req.params;
+        if(isNullValue(id))throw new ApiError('invalid input!' ,400)
         const getTeam = await team.findByPk(id);
-        console.log(getTeam.email != req.user.email)
         if(getTeam.email === req.user.email){
             await team.destroy({where:{id:id}});
             res.status(200).json({success:true});
@@ -73,66 +71,53 @@ exports.deleteTeam = async(req,res)=>{
             res.status(403).json({success:true});
         }
     }catch(err){
-        console.log(err)
-        res.status(500).json({success:false});
+        next(new ApiError(err.message ,err.statusCode))
     }
 }
 
-exports.addPlayer = async(req,res)=>{
+exports.addPlayer = async(req,res,next)=>{
     try {
-        const {email ,password ,teamId} = req.body;
+        let {name ,password ,teamId} = req.body;
+        if(isNullValue(name) || isNullValue(password)||isNullValue(teamId))throw new ApiError('invalid input!' ,400)
         const getTeamCount = await teamList.count({where:{teamId:teamId}});
         const getTeam = await team.findByPk(teamId);
-        console.log(getTeam.email , req.user.email);
-        if(getTeam.email === req.user.email){
-            if(getTeamCount < 11){
-                const getUser = await user.findOne({where:{email:email}});
-                // console.log('getUser',getUser)
-                if(getUser){
+        name = name.trim();
+        password = password.trim();
+        if(getTeam.captainName != req.user.name)throw new ApiError('You are not a captain!',400)
+            if(getTeamCount > 11)throw new ApiError('team is full!',400)
+                const getUser = await user.findOne({where:{name:name}});
+                if(!getUser)throw new ApiError('user not found!',404)
                     await hash.compareHash(password , getUser.password);
                     await teamList.create({
                         userId:getUser.id,
                         teamId:teamId
                     });
                     res.status(201).json({success:true,player:getUser});
-                }else{
-                    res.status(404).json({success:false,message:'user not found!'});
-                }
-            }else{
-                res.status(403).json({success:false,message:'team is full!'});
-            }
-        }else{
-            res.status(403).json({success:false,message:'you are not captain!'});
-        }  
+                
     } catch (err) {
-        console.log(err)
-        res.status(500).json({success:false ,message:err.message})
+        next(new ApiError(err.message ,err.statusCode))
+        
     }
 }
 
-exports.removePlayer = async(req ,res)=>{
+exports.removePlayer = async(req ,res,next)=>{
     try {
         const {id,teamId} = req.params;
+        if(isNullValue(id) || isNullValue(teamId))throw new ApiError('invalid input!' ,400)
         const getTeam = await team.findByPk(teamId);
         const getUser = await user.findByPk(id);
-        if(getUser.email != req.user.email){
-            if(getTeam.email === req.user.email){
-                    
-                await teamList.destroy({where:{
-                    userId:id,
-                    teamId:teamId
-                }});
-                res.status(201).json({success:true});
-                    
-            }else{
-                res.status(403).json({success:false,message:'you are not captain!'});
-            }  
-        }else{
-            res.status(403).json({success:false,message:'captain cannot be removed'});
-        }
+        if(getUser.name === req.user.name)throw new ApiError('captain cannot be removed!',400)
+        if(getTeam.captainName != req.user.name)throw new ApiError('You are not captain!',400)     
+        await teamList.destroy({where:{
+                userId:id,
+                teamId:teamId
+        }});
+        res.status(201).json({success:true});           
     } catch (err) {
-        console.log(err)
-        res.status(500).json({success:false ,message:err.message})
+        next(new ApiError(err.message ,err.statusCode))
     }
 }
 
+function isNullValue(value){
+    return value === ""?true :false;
+}
